backend: modified backend so that it is congruent with the OpenAPI specification of the next commit

+++ service/database/delete-comment.go
@@ -3,7 +3,16 @@ package database
 func (db *appdbimpl) Delete_comment(commentId uint64, photoId uint64) error {
 	// Deletes a comment from the database. if the comment was already deleted
 	// Implements DELETE /photos/{postID}/comments/
-	_, err := db.c.Exec(
+
+	isCID, _, _, err := db.IsCommentId(commentId)
+	if err != nil {
+		return err
+	}
+	if !isCID {
+		return nil
+	}
+
+	_, err = db.c.Exec(
 		`
 		BEGIN TRANSACTION;
 		DELETE FROM comments WHERE id = ?;
